+++ src/dg_cantor_pairing.cpp
@@ -3,7 +3,7 @@
 /// Cantor pairing function. Bijection two integers into one unique integer.
 /// @param x integer 1.
 /// @param y integer 2.
-int Cantor_pairing_fun(int x, int y){
+long long int Cantor_pairing_fun(long long int x, long long int y){
 	
 	return ((x + y) * (x + y + 1) / 2 + y);
 
@@ -15,10 +15,10 @@ int Cantor_pairing_fun(int x, int y){
 /// @param i x index. 
 /// @param j y index. 
 /// @param k level.  
-int Get_key_fun(int i, int j, int k){
+long long int Get_key_fun(int i, int j, int k){
  	
-	int key = Cantor_pairing_fun(i, j);
-	key = Cantor_pairing_fun(key, k);
+	long long int key = Cantor_pairing_fun((long long int)i, (long long int)j);
+	key = Cantor_pairing_fun(key, (long long int)k);
 
 	return key;
